[NABI-28] renamed test files and reorganized boilerplate
