+++ src/ts/Users/components/Registration.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import Typography from 'material-ui/Typography';
 import RegistrationForm from './RegistrationForm';
-import { UserState, Role } from '../../model';
+import { UserState, Role } from '../model';
 import { Redirect } from 'react-router-dom';
-import { InstructorState } from '../../../Instructors/model';
+import { InstructorState } from '../../Instructors/model';
 
 interface RegistrationProps {
   createUser: (user: UserState) => void;
